[#7] fix: Removed duplicate code
